refactor: update_and_delete_events

+++ src/services/event.service.ts
@@ -1,5 +1,9 @@
 import { event_model, event } from "../models/events.model";
 
+const isMyEvent = (userId: string, event: event): boolean => {
+    return event.organizer_id === userId; // Asegúrate de que tienes el campo organizerId en tu modelo de evento
+};
+
 class EventService {
     public async find_all_events(): Promise<event[]> {
         try{
@@ -26,16 +30,21 @@ class EventService {
         }
     }
 
-    public async update_event(id: string, event: event): Promise<event | null> {
+    public async update_event(id: string, event: event,user_id:string): Promise<event | null> {
         try{
+            if(!isMyEvent(user_id, event)) 
+            throw new Error('You are not the organizer of this event!');
             return await event_model.findByIdAndUpdate(id, event, {new:true});
         }catch(error){
             throw new Error(`A problem ocurred :( ${error}`);
         }
     }
 
-    public async delete_event(id: string): Promise<event | null> {
+    public async delete_event(id: string,user_id:string): Promise<event | null> {
         try{
+            const event_found = await event_model.findById(id);
+            if(event_found && !isMyEvent(user_id, event_found)) 
+                throw new Error('You are not the organizer of this event!');
             return await event_model.findByIdAndDelete(id);
         }catch(error){
             throw new Error(`A problem ocurred :( ${error}`);
